split amplitude estimation unit test into aqua and finance
